feat: implement useFlash composable for handling success messages in Designations page
